working on getting tsconfig right
eventually this will run in the browser, so split tsconfig

server tsconfig will be based on a modern version of node, incl.
modern node internals, but the browser stuff needs to not recognize
node internals. idk, just trying to get the right symbol resolution
in the right places, like, import.meta.dirname on server, if possible.